+++ CHANGELOG.md
@@ -52,3 +52,4 @@ Forget everything that was there before.
 * Cycles in Constraint Graph lead to a unified eq class.
 * Fix: Check for equality of bounds before reporting a conflict in exact bounds.
 * No longer discard reification errors.
+* Move TcKey management into ConstraintGraph, ditch VertexRef.